Introduction

AutoInteract

The main use of AutoInteract is to build widgets, extract signals from them, and apply them back to data from a live interface. These steps all maintain a similar structure to the original data so that it can all be automated.

Widgets

AutoInteract.make_widgets — Function.

make_widgets(x, label)

Widets are built recursively from any type, labelled using the dict or struct label, AxisArray names, etc.

They can then be edited before retrieving signals and building an interface.

source

Interface

AutoInteract.make_interface — Function.

make_interface(xs; box)

Interfaces are build recursively from any type containing widgets. This mostly means anything returned by make_widgets, but also after having any sub-components deleted custom components added:

source

Signals

AutoInteract.get_signals — Function.

get_signals(x)

Signals can be recursively retrieved from any type containing widgets.

source

AutoInteract.apply_signals! — Function.

apply_signals!(data, signal)

Signals can be recursively appied to data.

source

Plottables and plotting

AutoInteract also generates an interface for plotting arbitrary data in an interface. 'Plottables' are structured objects containing toggle-button widgets indicating if any data can be plotted. Mostly this means it inherits AbstractArray, or it is the last dimension of an AxisArray, with a grid of named toggle-buttons being made for the other dimensions. Signals from plottables can be plotted with plot_all().

AutoInteract.make_plottables — Function.

make_plottables(x, label)

Generate 'plotable' toggles recursively, with optional label.

source

AutoInteract.plot_all — Function.

plot_all(data, plottables, range, [label])

Plot all data for selected plottables, over the specified range.

source

AutoInteract.plot_all! — Function.

plot_all!(plots, data, plottables, range, label)

Recursive plotting methods for (almost) all types.

source

Misc

AutoInteract.spreadwidgets — Function.

spreadwidgets(widgets; cols = 6)

Create multiple columns to spread widgets accross a page

source